refactor: add ctor without limit
