+++ tests/Yaapii.Atoms.Tests/Scalar/HeadOfTest.cs
@@ -20,17 +20,12 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-using System;
-using System.Collections.Generic;
-using System.Text;
 using Xunit;
 using Yaapii.Atoms.Enumerable;
-using Yaapii.Atoms.List;
 using Yaapii.Atoms.Misc;
 using Yaapii.Atoms.Number;
-using Yaapii.Atoms.Scalar;
 
-namespace Yaapii.Atoms.Enumerable.Tests
+namespace Yaapii.Atoms.Scalar.Tests
 {
     public sealed class HeadOfTest
     {
